feat(QPage): Suggestion for not using QPageContainer (#10788)
see #10788
