+++ ui/src/components/page/QPageContainer.json
@@ -5,7 +5,7 @@
 
   "slots": {
     "default": {
-      "desc": "Encapsulates a QPage (either directly or through <router-view>)"
+      "desc": "depreciated"
     }
   }
 }
